Fix Show Current Player button

+++ src/app/player.h
@@ -8,6 +8,6 @@
 
 Player getPlayer(void *handle, bool skipIsValidCheck, uint64_t personAddress, uint64_t playerAddress);
 uint64_t getPersonAddressFromPlayerAddress(void *handle, uint64_t playerAddress);
-uint32_t getCurrentPersonUniqueId(const ProcessContext *processContext);
-Player getPlayerById(const ProcessContext *processContext, uint32_t uniqueId);
+uint64_t getCurrentPersonUniqueId(const ProcessContext *processContext);
+Player getPlayerById(const ProcessContext *processContext, uint64_t uniqueId);
 PositionWeights *getWeightsForPosition(PositionGrouped position);
